LayoutTests/fast/css/parsing-css-nth-child-of-3-expected.txt

Test the parsing of :nth-child(of) for querySelector and style.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test invalid selectors:
PASS document.querySelector(":nth-child(even of)") threw exception SyntaxError: ':nth-child(even of)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even of    )") threw exception SyntaxError: ':nth-child(even of    )' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(evenof .class)") threw exception SyntaxError: ':nth-child(evenof .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd of)") threw exception SyntaxError: ':nth-child(odd of)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd of    )") threw exception SyntaxError: ':nth-child(odd of    )' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(oddof .class)") threw exception SyntaxError: ':nth-child(oddof .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n of)") threw exception SyntaxError: ':nth-child(n of)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n of    )") threw exception SyntaxError: ':nth-child(n of    )' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(nof .class)") threw exception SyntaxError: ':nth-child(nof .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n of)") threw exception SyntaxError: ':nth-child(-n of)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n of    )") threw exception SyntaxError: ':nth-child(-n of    )' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-nof .class)") threw exception SyntaxError: ':nth-child(-nof .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 of)") threw exception SyntaxError: ':nth-child(3 of)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 of    )") threw exception SyntaxError: ':nth-child(3 of    )' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3of .class)") threw exception SyntaxError: ':nth-child(3of .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 of)") threw exception SyntaxError: ':nth-child(-3 of)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 of    )") threw exception SyntaxError: ':nth-child(-3 of    )' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3of .class)") threw exception SyntaxError: ':nth-child(-3of .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 of)") threw exception SyntaxError: ':nth-child(n+0 of)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 of    )") threw exception SyntaxError: ':nth-child(n+0 of    )' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0of .class)") threw exception SyntaxError: ':nth-child(n+0of .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 of)") threw exception SyntaxError: ':nth-child(n-0 of)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 of    )") threw exception SyntaxError: ':nth-child(n-0 of    )' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0of .class)") threw exception SyntaxError: ':nth-child(n-0of .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n of)") threw exception SyntaxError: ':nth-child(0n of)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n of    )") threw exception SyntaxError: ':nth-child(0n of    )' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0nof .class)") threw exception SyntaxError: ':nth-child(0nof .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 of)") threw exception SyntaxError: ':nth-child(3n+5 of)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 of    )") threw exception SyntaxError: ':nth-child(3n+5 of    )' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5of .class)") threw exception SyntaxError: ':nth-child(3n+5of .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 of)") threw exception SyntaxError: ':nth-child(-3n+5 of)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 of    )") threw exception SyntaxError: ':nth-child(-3n+5 of    )' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5of .class)") threw exception SyntaxError: ':nth-child(-3n+5of .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 of)") threw exception SyntaxError: ':nth-child(3n-5 of)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 of    )") threw exception SyntaxError: ':nth-child(3n-5 of    )' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5of .class)") threw exception SyntaxError: ':nth-child(3n-5of .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 of)") threw exception SyntaxError: ':nth-child(-3n-5 of)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 of    )") threw exception SyntaxError: ':nth-child(-3n-5 of    )' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5of .class)") threw exception SyntaxError: ':nth-child(-3n-5of .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even empty .class)") threw exception SyntaxError: ':nth-child(even empty .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even from .class)") threw exception SyntaxError: ':nth-child(even from .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even to .class)") threw exception SyntaxError: ':nth-child(even to .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even webkit .class)") threw exception SyntaxError: ':nth-child(even webkit .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd empty .class)") threw exception SyntaxError: ':nth-child(odd empty .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd from .class)") threw exception SyntaxError: ':nth-child(odd from .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd to .class)") threw exception SyntaxError: ':nth-child(odd to .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd webkit .class)") threw exception SyntaxError: ':nth-child(odd webkit .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n empty .class)") threw exception SyntaxError: ':nth-child(n empty .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n from .class)") threw exception SyntaxError: ':nth-child(n from .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n to .class)") threw exception SyntaxError: ':nth-child(n to .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n webkit .class)") threw exception SyntaxError: ':nth-child(n webkit .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n empty .class)") threw exception SyntaxError: ':nth-child(-n empty .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n from .class)") threw exception SyntaxError: ':nth-child(-n from .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n to .class)") threw exception SyntaxError: ':nth-child(-n to .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n webkit .class)") threw exception SyntaxError: ':nth-child(-n webkit .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 empty .class)") threw exception SyntaxError: ':nth-child(3 empty .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 from .class)") threw exception SyntaxError: ':nth-child(3 from .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 to .class)") threw exception SyntaxError: ':nth-child(3 to .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 webkit .class)") threw exception SyntaxError: ':nth-child(3 webkit .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 empty .class)") threw exception SyntaxError: ':nth-child(-3 empty .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 from .class)") threw exception SyntaxError: ':nth-child(-3 from .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 to .class)") threw exception SyntaxError: ':nth-child(-3 to .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 webkit .class)") threw exception SyntaxError: ':nth-child(-3 webkit .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 empty .class)") threw exception SyntaxError: ':nth-child(n+0 empty .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 from .class)") threw exception SyntaxError: ':nth-child(n+0 from .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 to .class)") threw exception SyntaxError: ':nth-child(n+0 to .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 webkit .class)") threw exception SyntaxError: ':nth-child(n+0 webkit .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 empty .class)") threw exception SyntaxError: ':nth-child(n-0 empty .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 from .class)") threw exception SyntaxError: ':nth-child(n-0 from .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 to .class)") threw exception SyntaxError: ':nth-child(n-0 to .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 webkit .class)") threw exception SyntaxError: ':nth-child(n-0 webkit .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n empty .class)") threw exception SyntaxError: ':nth-child(0n empty .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n from .class)") threw exception SyntaxError: ':nth-child(0n from .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n to .class)") threw exception SyntaxError: ':nth-child(0n to .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n webkit .class)") threw exception SyntaxError: ':nth-child(0n webkit .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 empty .class)") threw exception SyntaxError: ':nth-child(3n+5 empty .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 from .class)") threw exception SyntaxError: ':nth-child(3n+5 from .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 to .class)") threw exception SyntaxError: ':nth-child(3n+5 to .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 webkit .class)") threw exception SyntaxError: ':nth-child(3n+5 webkit .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 empty .class)") threw exception SyntaxError: ':nth-child(-3n+5 empty .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 from .class)") threw exception SyntaxError: ':nth-child(-3n+5 from .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 to .class)") threw exception SyntaxError: ':nth-child(-3n+5 to .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 webkit .class)") threw exception SyntaxError: ':nth-child(-3n+5 webkit .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 empty .class)") threw exception SyntaxError: ':nth-child(3n-5 empty .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 from .class)") threw exception SyntaxError: ':nth-child(3n-5 from .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 to .class)") threw exception SyntaxError: ':nth-child(3n-5 to .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 webkit .class)") threw exception SyntaxError: ':nth-child(3n-5 webkit .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 empty .class)") threw exception SyntaxError: ':nth-child(-3n-5 empty .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 from .class)") threw exception SyntaxError: ':nth-child(-3n-5 from .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 to .class)") threw exception SyntaxError: ':nth-child(-3n-5 to .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 webkit .class)") threw exception SyntaxError: ':nth-child(-3n-5 webkit .class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even of ::first-letter)") threw exception SyntaxError: ':nth-child(even of ::first-letter)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even of ::before)") threw exception SyntaxError: ':nth-child(even of ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even of ::after)") threw exception SyntaxError: ':nth-child(even of ::after)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even of ::-webkit-custom)") threw exception SyntaxError: ':nth-child(even of ::-webkit-custom)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even of .foo, ::before)") threw exception SyntaxError: ':nth-child(even of .foo, ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even of ::before, .foo)") threw exception SyntaxError: ':nth-child(even of ::before, .foo)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even of -webkit-any(::before, .foo))") threw exception SyntaxError: ':nth-child(even of -webkit-any(::before, .foo))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even of :nth-child(2n+1 of ::before))") threw exception SyntaxError: ':nth-child(even of :nth-child(2n+1 of ::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even of :not(::before))") threw exception SyntaxError: ':nth-child(even of :not(::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd of ::first-letter)") threw exception SyntaxError: ':nth-child(odd of ::first-letter)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd of ::before)") threw exception SyntaxError: ':nth-child(odd of ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd of ::after)") threw exception SyntaxError: ':nth-child(odd of ::after)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd of ::-webkit-custom)") threw exception SyntaxError: ':nth-child(odd of ::-webkit-custom)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd of .foo, ::before)") threw exception SyntaxError: ':nth-child(odd of .foo, ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd of ::before, .foo)") threw exception SyntaxError: ':nth-child(odd of ::before, .foo)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd of -webkit-any(::before, .foo))") threw exception SyntaxError: ':nth-child(odd of -webkit-any(::before, .foo))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd of :nth-child(2n+1 of ::before))") threw exception SyntaxError: ':nth-child(odd of :nth-child(2n+1 of ::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd of :not(::before))") threw exception SyntaxError: ':nth-child(odd of :not(::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n of ::first-letter)") threw exception SyntaxError: ':nth-child(n of ::first-letter)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n of ::before)") threw exception SyntaxError: ':nth-child(n of ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n of ::after)") threw exception SyntaxError: ':nth-child(n of ::after)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n of ::-webkit-custom)") threw exception SyntaxError: ':nth-child(n of ::-webkit-custom)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n of .foo, ::before)") threw exception SyntaxError: ':nth-child(n of .foo, ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n of ::before, .foo)") threw exception SyntaxError: ':nth-child(n of ::before, .foo)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n of -webkit-any(::before, .foo))") threw exception SyntaxError: ':nth-child(n of -webkit-any(::before, .foo))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n of :nth-child(2n+1 of ::before))") threw exception SyntaxError: ':nth-child(n of :nth-child(2n+1 of ::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n of :not(::before))") threw exception SyntaxError: ':nth-child(n of :not(::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n of ::first-letter)") threw exception SyntaxError: ':nth-child(-n of ::first-letter)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n of ::before)") threw exception SyntaxError: ':nth-child(-n of ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n of ::after)") threw exception SyntaxError: ':nth-child(-n of ::after)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n of ::-webkit-custom)") threw exception SyntaxError: ':nth-child(-n of ::-webkit-custom)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n of .foo, ::before)") threw exception SyntaxError: ':nth-child(-n of .foo, ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n of ::before, .foo)") threw exception SyntaxError: ':nth-child(-n of ::before, .foo)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n of -webkit-any(::before, .foo))") threw exception SyntaxError: ':nth-child(-n of -webkit-any(::before, .foo))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n of :nth-child(2n+1 of ::before))") threw exception SyntaxError: ':nth-child(-n of :nth-child(2n+1 of ::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n of :not(::before))") threw exception SyntaxError: ':nth-child(-n of :not(::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 of ::first-letter)") threw exception SyntaxError: ':nth-child(3 of ::first-letter)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 of ::before)") threw exception SyntaxError: ':nth-child(3 of ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 of ::after)") threw exception SyntaxError: ':nth-child(3 of ::after)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 of ::-webkit-custom)") threw exception SyntaxError: ':nth-child(3 of ::-webkit-custom)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 of .foo, ::before)") threw exception SyntaxError: ':nth-child(3 of .foo, ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 of ::before, .foo)") threw exception SyntaxError: ':nth-child(3 of ::before, .foo)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 of -webkit-any(::before, .foo))") threw exception SyntaxError: ':nth-child(3 of -webkit-any(::before, .foo))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 of :nth-child(2n+1 of ::before))") threw exception SyntaxError: ':nth-child(3 of :nth-child(2n+1 of ::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 of :not(::before))") threw exception SyntaxError: ':nth-child(3 of :not(::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 of ::first-letter)") threw exception SyntaxError: ':nth-child(-3 of ::first-letter)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 of ::before)") threw exception SyntaxError: ':nth-child(-3 of ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 of ::after)") threw exception SyntaxError: ':nth-child(-3 of ::after)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 of ::-webkit-custom)") threw exception SyntaxError: ':nth-child(-3 of ::-webkit-custom)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 of .foo, ::before)") threw exception SyntaxError: ':nth-child(-3 of .foo, ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 of ::before, .foo)") threw exception SyntaxError: ':nth-child(-3 of ::before, .foo)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 of -webkit-any(::before, .foo))") threw exception SyntaxError: ':nth-child(-3 of -webkit-any(::before, .foo))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 of :nth-child(2n+1 of ::before))") threw exception SyntaxError: ':nth-child(-3 of :nth-child(2n+1 of ::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 of :not(::before))") threw exception SyntaxError: ':nth-child(-3 of :not(::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 of ::first-letter)") threw exception SyntaxError: ':nth-child(n+0 of ::first-letter)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 of ::before)") threw exception SyntaxError: ':nth-child(n+0 of ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 of ::after)") threw exception SyntaxError: ':nth-child(n+0 of ::after)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 of ::-webkit-custom)") threw exception SyntaxError: ':nth-child(n+0 of ::-webkit-custom)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 of .foo, ::before)") threw exception SyntaxError: ':nth-child(n+0 of .foo, ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 of ::before, .foo)") threw exception SyntaxError: ':nth-child(n+0 of ::before, .foo)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 of -webkit-any(::before, .foo))") threw exception SyntaxError: ':nth-child(n+0 of -webkit-any(::before, .foo))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 of :nth-child(2n+1 of ::before))") threw exception SyntaxError: ':nth-child(n+0 of :nth-child(2n+1 of ::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 of :not(::before))") threw exception SyntaxError: ':nth-child(n+0 of :not(::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 of ::first-letter)") threw exception SyntaxError: ':nth-child(n-0 of ::first-letter)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 of ::before)") threw exception SyntaxError: ':nth-child(n-0 of ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 of ::after)") threw exception SyntaxError: ':nth-child(n-0 of ::after)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 of ::-webkit-custom)") threw exception SyntaxError: ':nth-child(n-0 of ::-webkit-custom)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 of .foo, ::before)") threw exception SyntaxError: ':nth-child(n-0 of .foo, ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 of ::before, .foo)") threw exception SyntaxError: ':nth-child(n-0 of ::before, .foo)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 of -webkit-any(::before, .foo))") threw exception SyntaxError: ':nth-child(n-0 of -webkit-any(::before, .foo))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 of :nth-child(2n+1 of ::before))") threw exception SyntaxError: ':nth-child(n-0 of :nth-child(2n+1 of ::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 of :not(::before))") threw exception SyntaxError: ':nth-child(n-0 of :not(::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n of ::first-letter)") threw exception SyntaxError: ':nth-child(0n of ::first-letter)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n of ::before)") threw exception SyntaxError: ':nth-child(0n of ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n of ::after)") threw exception SyntaxError: ':nth-child(0n of ::after)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n of ::-webkit-custom)") threw exception SyntaxError: ':nth-child(0n of ::-webkit-custom)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n of .foo, ::before)") threw exception SyntaxError: ':nth-child(0n of .foo, ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n of ::before, .foo)") threw exception SyntaxError: ':nth-child(0n of ::before, .foo)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n of -webkit-any(::before, .foo))") threw exception SyntaxError: ':nth-child(0n of -webkit-any(::before, .foo))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n of :nth-child(2n+1 of ::before))") threw exception SyntaxError: ':nth-child(0n of :nth-child(2n+1 of ::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n of :not(::before))") threw exception SyntaxError: ':nth-child(0n of :not(::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 of ::first-letter)") threw exception SyntaxError: ':nth-child(3n+5 of ::first-letter)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 of ::before)") threw exception SyntaxError: ':nth-child(3n+5 of ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 of ::after)") threw exception SyntaxError: ':nth-child(3n+5 of ::after)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 of ::-webkit-custom)") threw exception SyntaxError: ':nth-child(3n+5 of ::-webkit-custom)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 of .foo, ::before)") threw exception SyntaxError: ':nth-child(3n+5 of .foo, ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 of ::before, .foo)") threw exception SyntaxError: ':nth-child(3n+5 of ::before, .foo)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 of -webkit-any(::before, .foo))") threw exception SyntaxError: ':nth-child(3n+5 of -webkit-any(::before, .foo))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 of :nth-child(2n+1 of ::before))") threw exception SyntaxError: ':nth-child(3n+5 of :nth-child(2n+1 of ::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 of :not(::before))") threw exception SyntaxError: ':nth-child(3n+5 of :not(::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 of ::first-letter)") threw exception SyntaxError: ':nth-child(-3n+5 of ::first-letter)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 of ::before)") threw exception SyntaxError: ':nth-child(-3n+5 of ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 of ::after)") threw exception SyntaxError: ':nth-child(-3n+5 of ::after)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 of ::-webkit-custom)") threw exception SyntaxError: ':nth-child(-3n+5 of ::-webkit-custom)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 of .foo, ::before)") threw exception SyntaxError: ':nth-child(-3n+5 of .foo, ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 of ::before, .foo)") threw exception SyntaxError: ':nth-child(-3n+5 of ::before, .foo)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 of -webkit-any(::before, .foo))") threw exception SyntaxError: ':nth-child(-3n+5 of -webkit-any(::before, .foo))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 of :nth-child(2n+1 of ::before))") threw exception SyntaxError: ':nth-child(-3n+5 of :nth-child(2n+1 of ::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 of :not(::before))") threw exception SyntaxError: ':nth-child(-3n+5 of :not(::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 of ::first-letter)") threw exception SyntaxError: ':nth-child(3n-5 of ::first-letter)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 of ::before)") threw exception SyntaxError: ':nth-child(3n-5 of ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 of ::after)") threw exception SyntaxError: ':nth-child(3n-5 of ::after)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 of ::-webkit-custom)") threw exception SyntaxError: ':nth-child(3n-5 of ::-webkit-custom)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 of .foo, ::before)") threw exception SyntaxError: ':nth-child(3n-5 of .foo, ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 of ::before, .foo)") threw exception SyntaxError: ':nth-child(3n-5 of ::before, .foo)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 of -webkit-any(::before, .foo))") threw exception SyntaxError: ':nth-child(3n-5 of -webkit-any(::before, .foo))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 of :nth-child(2n+1 of ::before))") threw exception SyntaxError: ':nth-child(3n-5 of :nth-child(2n+1 of ::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 of :not(::before))") threw exception SyntaxError: ':nth-child(3n-5 of :not(::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 of ::first-letter)") threw exception SyntaxError: ':nth-child(-3n-5 of ::first-letter)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 of ::before)") threw exception SyntaxError: ':nth-child(-3n-5 of ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 of ::after)") threw exception SyntaxError: ':nth-child(-3n-5 of ::after)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 of ::-webkit-custom)") threw exception SyntaxError: ':nth-child(-3n-5 of ::-webkit-custom)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 of .foo, ::before)") threw exception SyntaxError: ':nth-child(-3n-5 of .foo, ::before)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 of ::before, .foo)") threw exception SyntaxError: ':nth-child(-3n-5 of ::before, .foo)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 of -webkit-any(::before, .foo))") threw exception SyntaxError: ':nth-child(-3n-5 of -webkit-any(::before, .foo))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 of :nth-child(2n+1 of ::before))") threw exception SyntaxError: ':nth-child(-3n-5 of :nth-child(2n+1 of ::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 of :not(::before))") threw exception SyntaxError: ':nth-child(-3n-5 of :not(::before))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even of .123class)") threw exception SyntaxError: ':nth-child(even of .123class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even of #123id)") threw exception SyntaxError: ':nth-child(even of #123id)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even of [])") threw exception SyntaxError: ':nth-child(even of [])' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even of ())") threw exception SyntaxError: ':nth-child(even of ())' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even of ))") threw exception SyntaxError: ':nth-child(even of ))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even of {})") threw exception SyntaxError: ':nth-child(even of {})' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(even of })") threw exception SyntaxError: ':nth-child(even of })' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd of .123class)") threw exception SyntaxError: ':nth-child(odd of .123class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd of #123id)") threw exception SyntaxError: ':nth-child(odd of #123id)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd of [])") threw exception SyntaxError: ':nth-child(odd of [])' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd of ())") threw exception SyntaxError: ':nth-child(odd of ())' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd of ))") threw exception SyntaxError: ':nth-child(odd of ))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd of {})") threw exception SyntaxError: ':nth-child(odd of {})' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(odd of })") threw exception SyntaxError: ':nth-child(odd of })' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n of .123class)") threw exception SyntaxError: ':nth-child(n of .123class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n of #123id)") threw exception SyntaxError: ':nth-child(n of #123id)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n of [])") threw exception SyntaxError: ':nth-child(n of [])' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n of ())") threw exception SyntaxError: ':nth-child(n of ())' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n of ))") threw exception SyntaxError: ':nth-child(n of ))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n of {})") threw exception SyntaxError: ':nth-child(n of {})' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n of })") threw exception SyntaxError: ':nth-child(n of })' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n of .123class)") threw exception SyntaxError: ':nth-child(-n of .123class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n of #123id)") threw exception SyntaxError: ':nth-child(-n of #123id)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n of [])") threw exception SyntaxError: ':nth-child(-n of [])' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n of ())") threw exception SyntaxError: ':nth-child(-n of ())' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n of ))") threw exception SyntaxError: ':nth-child(-n of ))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n of {})") threw exception SyntaxError: ':nth-child(-n of {})' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-n of })") threw exception SyntaxError: ':nth-child(-n of })' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 of .123class)") threw exception SyntaxError: ':nth-child(3 of .123class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 of #123id)") threw exception SyntaxError: ':nth-child(3 of #123id)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 of [])") threw exception SyntaxError: ':nth-child(3 of [])' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 of ())") threw exception SyntaxError: ':nth-child(3 of ())' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 of ))") threw exception SyntaxError: ':nth-child(3 of ))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 of {})") threw exception SyntaxError: ':nth-child(3 of {})' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3 of })") threw exception SyntaxError: ':nth-child(3 of })' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 of .123class)") threw exception SyntaxError: ':nth-child(-3 of .123class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 of #123id)") threw exception SyntaxError: ':nth-child(-3 of #123id)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 of [])") threw exception SyntaxError: ':nth-child(-3 of [])' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 of ())") threw exception SyntaxError: ':nth-child(-3 of ())' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 of ))") threw exception SyntaxError: ':nth-child(-3 of ))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 of {})") threw exception SyntaxError: ':nth-child(-3 of {})' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3 of })") threw exception SyntaxError: ':nth-child(-3 of })' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 of .123class)") threw exception SyntaxError: ':nth-child(n+0 of .123class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 of #123id)") threw exception SyntaxError: ':nth-child(n+0 of #123id)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 of [])") threw exception SyntaxError: ':nth-child(n+0 of [])' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 of ())") threw exception SyntaxError: ':nth-child(n+0 of ())' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 of ))") threw exception SyntaxError: ':nth-child(n+0 of ))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 of {})") threw exception SyntaxError: ':nth-child(n+0 of {})' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n+0 of })") threw exception SyntaxError: ':nth-child(n+0 of })' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 of .123class)") threw exception SyntaxError: ':nth-child(n-0 of .123class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 of #123id)") threw exception SyntaxError: ':nth-child(n-0 of #123id)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 of [])") threw exception SyntaxError: ':nth-child(n-0 of [])' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 of ())") threw exception SyntaxError: ':nth-child(n-0 of ())' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 of ))") threw exception SyntaxError: ':nth-child(n-0 of ))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 of {})") threw exception SyntaxError: ':nth-child(n-0 of {})' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(n-0 of })") threw exception SyntaxError: ':nth-child(n-0 of })' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n of .123class)") threw exception SyntaxError: ':nth-child(0n of .123class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n of #123id)") threw exception SyntaxError: ':nth-child(0n of #123id)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n of [])") threw exception SyntaxError: ':nth-child(0n of [])' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n of ())") threw exception SyntaxError: ':nth-child(0n of ())' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n of ))") threw exception SyntaxError: ':nth-child(0n of ))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n of {})") threw exception SyntaxError: ':nth-child(0n of {})' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(0n of })") threw exception SyntaxError: ':nth-child(0n of })' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 of .123class)") threw exception SyntaxError: ':nth-child(3n+5 of .123class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 of #123id)") threw exception SyntaxError: ':nth-child(3n+5 of #123id)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 of [])") threw exception SyntaxError: ':nth-child(3n+5 of [])' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 of ())") threw exception SyntaxError: ':nth-child(3n+5 of ())' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 of ))") threw exception SyntaxError: ':nth-child(3n+5 of ))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 of {})") threw exception SyntaxError: ':nth-child(3n+5 of {})' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n+5 of })") threw exception SyntaxError: ':nth-child(3n+5 of })' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 of .123class)") threw exception SyntaxError: ':nth-child(-3n+5 of .123class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 of #123id)") threw exception SyntaxError: ':nth-child(-3n+5 of #123id)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 of [])") threw exception SyntaxError: ':nth-child(-3n+5 of [])' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 of ())") threw exception SyntaxError: ':nth-child(-3n+5 of ())' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 of ))") threw exception SyntaxError: ':nth-child(-3n+5 of ))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 of {})") threw exception SyntaxError: ':nth-child(-3n+5 of {})' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n+5 of })") threw exception SyntaxError: ':nth-child(-3n+5 of })' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 of .123class)") threw exception SyntaxError: ':nth-child(3n-5 of .123class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 of #123id)") threw exception SyntaxError: ':nth-child(3n-5 of #123id)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 of [])") threw exception SyntaxError: ':nth-child(3n-5 of [])' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 of ())") threw exception SyntaxError: ':nth-child(3n-5 of ())' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 of ))") threw exception SyntaxError: ':nth-child(3n-5 of ))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 of {})") threw exception SyntaxError: ':nth-child(3n-5 of {})' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(3n-5 of })") threw exception SyntaxError: ':nth-child(3n-5 of })' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 of .123class)") threw exception SyntaxError: ':nth-child(-3n-5 of .123class)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 of #123id)") threw exception SyntaxError: ':nth-child(-3n-5 of #123id)' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 of [])") threw exception SyntaxError: ':nth-child(-3n-5 of [])' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 of ())") threw exception SyntaxError: ':nth-child(-3n-5 of ())' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 of ))") threw exception SyntaxError: ':nth-child(-3n-5 of ))' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 of {})") threw exception SyntaxError: ':nth-child(-3n-5 of {})' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":nth-child(-3n-5 of })") threw exception SyntaxError: ':nth-child(-3n-5 of })' is not a valid selector..
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS successfullyParsed is true

TEST COMPLETE
